% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainLassoLDATraining.R
\name{bootstrap_scGPS}
\alias{bootstrap_scGPS}
\title{BootStrap}
\usage{
bootstrap_scGPS(nboots = 1, mixedpop1 = mixedpop1, mixedpop2 = mixedpop2,
  c_selectID, listData = list())
}
\arguments{
\item{nboots}{a number specifying how many bootstraps to be run}

\item{mixedpop1}{is a \linkS4class{SingleCellExperiment} object from the train mixed population}

\item{mixedpop2}{is a \linkS4class{SingleCellExperiment} object from the target mixed population
of importance, e.g. differentially expressed genes that are most significant}

\item{listData}{is a \code{list} object, which contains trained results for the first mixed population}
}
\value{
a \code{list} with prediction results written in to the index \code{out_idx}
}
\description{
Predict a new mixed population after training the model for a subpopulation in the first mixed population
}
\examples{
c_selectID<-1
out_idx<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, GeneMetadata = day2$dat2geneInfo,
                     CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts, GeneMetadata = day5$dat5geneInfo,
                     CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
test <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1, mixedpop2 = mixedpop2, c_selectID, listData =list())
names(test)
test$LassoPredict
test$LDAPredict
}
\author{
Quan Nguyen, 2017-11-25
}
