% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainLassoLDATraining.R
\name{bootstrap_scGPS_parallel}
\alias{bootstrap_scGPS_parallel}
\title{BootStrap runs for both scGPS training and prediction with parallel option}
\usage{
bootstrap_scGPS_parallel(ncores = 4, nboots = 1, genes = genes,
  mixedpop1 = mixedpop1, mixedpop2 = mixedpop2, c_selectID,
  listData = list())
}
\arguments{
\item{ncores}{a number specifying how many cpus to be used for running}

\item{nboots}{a number specifying how many bootstraps to be run}

\item{genes}{a gene list to build the model}

\item{mixedpop1}{a \linkS4class{SingleCellExperiment} object from a mixed population for training}

\item{mixedpop2}{a \linkS4class{SingleCellExperiment} object from a target mixed population for prediction}

\item{listData}{a \code{list} object, which contains trained results for the first mixed population}
}
\value{
a \code{list} with prediction results written in to the index \code{out_idx}
}
\description{
same as bootstrap_scGPS, but with an multicore option
}
\author{
Quan Nguyen, 2017-11-25
}
