% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcDist}
\alias{calcDist}
\title{Compute Euclidean distance matrix by rows}
\usage{
calcDist(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
a distance matrix
}
\description{
Compute Euclidean distance matrix by rows
}
\examples{
mat_test <-matrix(rnbinom(1000,mu=0.01, size=10),nrow=1000)
calcDist(mat_test)

}
