% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{CIDR_scGPS}
\alias{CIDR_scGPS}
\title{CIDR}
\usage{
CIDR_scGPS(expression.matrix = NULL, ngenes = 1500, scaling = TRUE,
  theta = 0.5, perplexity = 30)
}
\arguments{
\item{expression_matrix}{An expression matrix, with genes in rows}
}
\value{
a CIDR reduced matrix for the top 20 components
}
\description{
calculate CIDR using top variable genes
}
\examples{
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, GeneMetadata = day2$dat2geneInfo,
                    CellMetadata = day2$dat2_clusters)
t <-CIDR_scGPS(expression.matrix=assay(mixedpop1))

}
