% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{clustering_scGPS}
\alias{clustering_scGPS}
\title{Iterative HC clustering}
\usage{
clustering_scGPS(object = NULL, windows = seq(0.025:1, by = 0.025),
  remove_outlier = c(0))
}
\arguments{
\item{remove_outlier}{a vector containing IDs for clusters to be removed
the default vector contains 0, as 0 is the cluster with singletons}

\item{mixedpop1}{is a \linkS4class{SingleCellExperiment} object from the
train mixed population}
}
\value{
a \code{matrix} with Eucleadean distance used for clustreting
}
\description{
performs 40 clustering runs
}
\author{
Quan Nguyen, 2017-11-25
}
