% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{CORE_scGPS}
\alias{CORE_scGPS}
\title{Main clustering CORE V2.0}
\usage{
CORE_scGPS(mixedpop = NULL, windows = seq(0.025:1, by = 0.025),
  remove_outlier = c(0))
}
\arguments{
\item{mixedpop}{is a \linkS4class{SingleCellExperiment} object from the train
mixed population}

\item{windows}{a numeric specifying the number of windows to test}

\item{remove_outlier}{a vector containing IDs for clusters to be removed
the default vector contains 0, as 0 is the cluster with singletons.}
}
\value{
a \code{list} with clustering results of all iterations, and a selected
optimal resolution
}
\description{
CORE is an algorithm to generate reproduciable clustering,
CORE is first implemented in ascend R package. Here, CORE V2.0 introduces several new
functionalities, including three key feature:
fast (and more memory efficient) implementation with C++ and paralellisation options allowing clustering
of hundreds of thousands of cells (ongoing development), outlier revomal important if singletons
exist (done), a number of dimensionality reduction methods including the imputation
implementation (CIDR) for confirming clustering results (done), and an option
to select the number of optimisation tree height windows for increasing resolution
}
\examples{
day5 <- sample2
mixedpop2 <-NewscGPS_SME(ExpressionMatrix = day5$dat5_counts, GeneMetadata = day5$dat5geneInfo,
                         CellMetadata = day5$dat5_clusters)
test <- CORE_scGPS(mixedpop2, remove_outlier = c(1))
}
\author{
Quan Nguyen, 2017-11-25
}
