% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering_bagging.R
\name{CORE_scGPS_bagging}
\alias{CORE_scGPS_bagging}
\title{Main clustering CORE V2.0 with bagging analysis}
\usage{
CORE_scGPS_bagging(mixedpop = NULL, bagging_run = 10,
  subsample_proportion = 0.8, windows = seq(0.025:1, by = 0.025),
  remove_outlier = c(0), nRounds = 1, PCA = FALSE, nPCs = 20,
  ngenes = 1500)
}
\arguments{
\item{windows}{a numeric specifying the number of windows to test}

\item{remove_outlier}{a vector containing IDs for clusters to be removed
the default vector contains 0, as 0 is the cluster with singletons.}

\item{PCA}{logical specifying if PCA is used before calculating distance matrix}
}
\value{
a \code{list} with clustering results of all iterations, and a selected
optimal resolution
}
\description{
CORE is an algorithm to generate reproduciable clustering,
CORE is first implemented in ascend R package. Here, CORE V2.0 uses bagging analysis
to find a stable clustering result and detect rare clusters mixed population
}
\examples{
day5 <- sample2
cellnames<-colnames(day5$dat5_counts)
cluster <-day5$dat5_clusters
cellnames <- data.frame("cluster" = cluster, "cellBarcodes" = cellnames)
#day5$dat5_counts needs to be in a matrix format
mixedpop2 <-NewscGPS_SME(ExpressionMatrix = day5$dat5_counts, GeneMetadata = day5$dat5geneInfo,
                         CellMetadata = day5$dat5_clusters)
test <- CORE_scGPS_bagging(mixedpop2, remove_outlier = c(1), PCA=FALSE, bagging_run = 20, subsample_proportion = .8)
plot_CORE(test$tree, list_clusters = test$clustering_param)

}
\author{
Quan Nguyen, 2018-05-11
}
