% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{CORE_Subcluster_scGPS}
\alias{CORE_Subcluster_scGPS}
\title{Subclustering (optional) after running CORE}
\usage{
CORE_Subcluster_scGPS(mixedpop = NULL, windows = seq(0.025:1, by = 0.025),
  remove_other_clusters = c(0))
}
\arguments{
\item{mixedpop}{is a \linkS4class{SingleCellExperiment} object from the train
mixed population}

\item{windows}{a numeric specifying the number of windows to test}

\item{remove_other_clusters}{a vector containing IDs for clusters to be removed,
leaving remaining clusters for reclustering}
}
\value{
a \code{list} with clustering results of all iterations, and a selected
optimal resolution
}
\description{
CORE_Subcluster_scGPS allows re-cluster the CORE clustering result
}
\examples{
day5 <- sample2
mixedpop2 <-NewscGPS_SME(ExpressionMatrix = day5$dat5_counts, GeneMetadata = day5$dat5geneInfo,
                         CellMetadata = day5$dat5_clusters)
test <- CORE_scGPS(mixedpop2, remove_other_clusters = c(1))
}
\author{
Quan Nguyen, 2017-11-25
}
