% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_2_cardio_cell_sample.R
\docType{data}
\name{day_2_cardio_cell_sample}
\alias{day_2_cardio_cell_sample}
\title{One of the two example single-cell count matrices to be used
for training \pkg{scGPS} model}
\format{a list instance, containing a count matrix and a vector with 
clustering information}
\source{
Dr Joseph Powell's laboratory, IMB, UQ
}
\usage{
day_2_cardio_cell_sample
}
\value{
a list, with the name day2
}
\description{
The count data set contains counts for 16990 genes for 590 cells
randomly subsampled from day-2 cardio-differentiation population.
The vector of clustering information contains corresponding to cells in the 
count matrix
}
\author{
Quan Nguyen, 2017-11-25
}
\keyword{datasets}
