% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_5_cardio_cell_sample.R
\docType{data}
\name{day_5_cardio_cell_sample}
\alias{day_5_cardio_cell_sample}
\title{One of the two example single-cell count matrices to be used
for \pkg{scGPS} prediction}
\format{
a list instance, containing a count matrix and a vector with 
clustering information.
}
\source{
Dr Joseph Powell's laboratory, IMB, UQ
}
\usage{
day_5_cardio_cell_sample
}
\value{
a list, with the name day5
}
\description{
The count data set contains counts for 17402 genes for 983 cells
(1 row per gene) randomly subsampled from day-5 cardio-differentiation 
population
The vector of clustering information contains corresponding to cells in the
count matrix
}
\author{
Quan Nguyen, 2017-11-25
}
\keyword{datasets}
