% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{distance_scGPS}
\alias{distance_scGPS}
\title{build disance matrice V2.0}
\usage{
distance_scGPS(object = NULL)
}
\arguments{
\item{mixedpop1}{is a \linkS4class{SingleCellExperiment} object from the train mixed population}
}
\value{
a \code{matrix} with Eucleadean distance used for clustreting
}
\description{
Calculate distance matrix and perform hclust
}
\author{
Quan Nguyen, 2017-11-25
}
