% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distvec}
\alias{distvec}
\title{Compute Distance between two vectors}
\usage{
distvec(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
a numeric distance
}
\description{
Compute Distance between two vectors
}
\examples{
x <-matrix(rnbinom(1000,mu=0.01, size=10),nrow=1000)
x <- x[1,]
y <-matrix(rnbinom(1000,mu=0.01, size=10),nrow=1000)
y <- y[1,]
distvec(x, y)

}
