% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{FindOptimalStability}
\alias{FindOptimalStability}
\title{Find the optimal cluster}
\usage{
FindOptimalStability(list_clusters, run_RandIdx)
}
\arguments{
\item{list_clusters}{is a \code{list} object containing 40 clustering results}

\item{run_RandIdx}{is a \code{data frame} object from iterative clustering runs}
}
\value{
a \code{list} with optimal stability, cluster count and summary stats
}
\description{
Find the optimal cluster
}
\author{
Quan Nguyen, 2017-11-25
}
