% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{FindStability}
\alias{FindStability}
\title{Calculate stability index}
\usage{
FindStability(list_clusters = NULL, cluster_ref = NULL)
}
\arguments{
\item{list_clusters}{is a object from the iterative clustering runs}

\item{cluster_ref}{is a object from the reference cluster}
}
\value{
a \code{data frame} with stability scores and randIndex results
}
\description{
Calculate stability index
}
\author{
Quan Nguyen, 2017-11-25
}
