% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneList.R
\docType{data}
\name{GeneList}
\alias{GeneList}
\title{Input gene list for training \pkg{scGPS}, e.g. differentially 
expressed genes}
\format{a list instance, containing a count matrix and a vector with 
clustering information.}
\source{
Dr Joseph Powell's laboratory, IMB, UQ
}
\usage{
GeneList
}
\value{
a vector containing gene symbols
}
\description{
Genes should be ordered from most to least important genes
(1 row per gene)
}
\author{
Quan Nguyen, 2017-11-25
}
\keyword{datasets}
