% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{PCA_scGPS}
\alias{PCA_scGPS}
\title{PCA}
\usage{
PCA_scGPS(expression.matrix = NULL, ngenes = 1500, scaling = TRUE,
  npcs = 50)
}
\arguments{
\item{expression_matrix}{An expression matrix, with genes in rows}
}
\value{
a list containing PCA results and variance explained
}
\description{
Select top variable genes and perform prcomp
}
\examples{
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, GeneMetadata = day2$dat2geneInfo,
                    CellMetadata = day2$dat2_clusters)
t <-PCA_scGPS(expression.matrix=assay(mixedpop1))

}
