% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{plot_CORE}
\alias{plot_CORE}
\title{Plot dendrogram tree for CORE result}
\usage{
plot_CORE(original.tree, list_clusters = NULL)
}
\description{
This function plots CORE and all clustering results underneath
}
