% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainLassoLDATraining.R
\name{predicting_scGPS}
\alias{predicting_scGPS}
\title{Main prediction}
\usage{
predicting_scGPS(listData = NULL, mixedpop2 = NULL, out_idx)
}
\arguments{
\item{listData}{is a \code{list} object, which contains trained results for the first mixed population}

\item{mixedpop2}{is a \linkS4class{SingleCellExperiment} object from the target mixed population
of importance, e.g. differentially expressed genes that are most significant}

\item{out_idx}{a number for index of the list}
}
\value{
a \code{list} with prediction results written in to the index \code{out_idx}
}
\description{
Predict a new mixed population after training the model for a subpopulation in the first mixed population
}
\examples{
c_selectID<-1
out_idx<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, GeneMetadata = day2$dat2geneInfo,
                     CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts, GeneMetadata = day5$dat5geneInfo,
                     CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
listData  <- training_scGPS(genes, mixedpop1 = mixedpop1, mixedpop2 = mixedpop2, c_selectID, listData =list(), out_idx=out_idx)
listData  <- predicting_scGPS(listData =listData,  mixedpop2 = mixedpop2, out_idx=out_idx)
}
\author{
Quan Nguyen, 2017-11-25
}
