% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PrinComp_cpp}
\alias{PrinComp_cpp}
\title{Principal component analysis}
\usage{
PrinComp_cpp(X)
}
\arguments{
\item{X}{an R matrix (expression matrix), rows are genes, columns are cells}
}
\description{
This function provides significant speed gain if the input matrix
is big
}
