% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{randIndex}
\alias{randIndex}
\title{Calculate rand index}
\usage{
randIndex(tab, adjust = TRUE)
}
\arguments{
\item{tab}{a table containing different clustering results in rows}

\item{adjust}{a logical of whether to use the adjusted rand index}
}
\value{
a randIndex value
}
\description{
Comparing clustering results Function for calculating randindex
(adapted from the function by Steve Horvath and Luohua Jiang, UCLA, 2003)
}
\examples{
day5 <- sample2
mixedpop2 <-NewscGPS_SME(ExpressionMatrix = day5$dat5_counts,
GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
cluster_all <-clustering_scGPS(object=mixedpop2)

randIndex(table(unlist(cluster_all$list_clusters[[1]]), 
cluster_all$cluster_ref))

}
\author{
Quan Nguyen and Michael Thompson, 2018-05-11
}
