% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_Eucl_distance_NotPar}
\alias{rcpp_Eucl_distance_NotPar}
\title{Function to calculate Eucledean distance matrix without parallelisation}
\usage{
rcpp_Eucl_distance_NotPar(mat)
}
\arguments{
\item{X}{an R matrix (expression matrix), with cells in rows and genes in columns}
}
\description{
Function to calculate Eucledean distance matrix without parallelisation
}
