% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_distance}
\alias{rcpp_parallel_distance}
\title{Function to calculate Eucledean distance matrix with paralleled C++}
\usage{
rcpp_parallel_distance(mat)
}
\arguments{
\item{X}{an R matrix (expression matrix), with cells in rows and genes in columns}
}
\description{
Function to calculate Eucledean distance matrix with paralleled C++
}
