% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset_cpp}
\alias{subset_cpp}
\title{Subset a matrix}
\usage{
subset_cpp(m1in, rowidx_in, colidx_in)
}
\arguments{
\item{X}{an R matrix (expression matrix)}
}
\value{
a subsetted matrix
}
\description{
Subset a matrix
}
\examples{
mat_test <-matrix(rnbinom(1000000,mu=0.01, size=10),nrow=100)
subset_mat <- subset_cpp(mat_test, rowidx_in=c(1:10), colidx_in=c(100:500))
dim(subset_mat)
}
