% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{topvar_scGPS}
\alias{topvar_scGPS}
\title{select top variable genes}
\usage{
topvar_scGPS(expression.matrix = NULL, ngenes = 1500)
}
\arguments{
\item{expression.matrix}{is a matrix with genes in rows and cells in columns}
}
\description{
subset a matrix by top variable genes
}
