% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tp_cpp}
\alias{tp_cpp}
\title{Transpose a matrix}
\usage{
tp_cpp(X)
}
\arguments{
\item{X}{an R matrix (expression matrix)}
}
\value{
a transposed matrix
}
\description{
Transpose a matrix
}
\examples{
mat_test <-matrix(rnbinom(1000000,mu=0.01, size=10),nrow=100)
tp_mat <- tp_cpp(mat_test)
}
