% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_gene_sample.R
\docType{data}
\name{training_gene_sample}
\alias{training_gene_sample}
\title{Input gene list for training \pkg{scGPS}, e.g. differentially 
expressed genes}
\format{
a list instance, containing a count matrix and a vector with 
clustering information.
}
\source{
Dr Joseph Powell's laboratory, IMB, UQ
}
\usage{
training_gene_sample
}
\value{
a vector containing gene symbols
}
\description{
Genes should be ordered from most to least important genes
(1 row per gene)
}
\author{
Quan Nguyen, 2017-11-25
}
\keyword{datasets}
