% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainLassoLDATraining.R
\name{training_scGPS}
\alias{training_scGPS}
\title{Main training}
\usage{
training_scGPS(genes, mixedpop1 = NULL, mixedpop2 = NULL, c_selectID,
  listData = list(), out_idx)
}
\arguments{
\item{genes}{a vector of gene names (for LASSO shrinkage); gene symbols must be
in the same format with gene names in subpop2; #' genes are listed by the order
of importance, e.g. differentially expressed genes that are most significant}

\item{mixedpop1}{is a \linkS4class{SingleCellExperiment} object from the train mixed population}

\item{mixedpop2}{is a \linkS4class{SingleCellExperiment} object from the target mixed population}

\item{c_selectID}{a selected number to specify which subpop to be used for training}

\item{out_idx}{a number for index of the list}
}
\value{
a \code{list} with prediction results written in to the index \code{out_idx}
}
\description{
Training LASSO and LDA
}
\examples{

c_selectID<-1
out_idx<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, GeneMetadata = day2$dat2geneInfo,
                     CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts, GeneMetadata = day5$dat5geneInfo,
                     CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
listData  <- training_scGPS(genes, mixedpop1 = mixedpop1, mixedpop2 = mixedpop2, c_selectID, listData =list(), out_idx=out_idx)
day2 <- sample1

}
\author{
Quan Nguyen, 2017-11-25
}
