% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_cpp}
\alias{var_cpp}
\title{Calculate variance}
\usage{
var_cpp(x, bias = TRUE)
}
\arguments{
\item{x}{a vector of gene expression.}

\item{bias}{degree of freedom}
}
\description{
Calculate variance
}
