% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varRcpp}
\alias{varRcpp}
\title{Calculate variance}
\usage{
varRcpp(x, bias = TRUE)
}
\arguments{
\item{x}{a vector of gene expression.}

\item{bias}{degree of freedom}
}
\description{
Calculate variance
}
