% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{annotate_scGPS}
\alias{annotate_scGPS}
\title{annotate_scGPS functionally annotates the identified clusters}
\usage{
annotate_scGPS(DEgeneList, pvalueCutoff = 0.05, gene_symbol = TRUE,
  output_filename = "PathwayEnrichment.xlsx", output_path = NULL)
}
\arguments{
\item{DEgeneList}{is a vector of gene symbols, convertable to ENTREZID}

\item{pvalueCutoff}{is a numeric of the cutoff p value}

\item{gene_symbol}{-------}

\item{output_filename}{is a string of the filename to save the spreadsheet to}

\item{output_path}{is a string of the path to the location to save the spreadsheet}
}
\value{
write enrichment test output to a file and an enrichment test object for plotting
}
\description{
often we need to label clusters with unique biological characters.
One of the common approach to annotate a cluster is to perform functional enrichment
analysis. The annotate_scGPS implements ReactomePA and clusterProfiler for this analysis
type in R. The function require installation of several databases as described below.
}
\examples{
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, GeneMetadata = day2$dat2geneInfo,
                     CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts, GeneMetadata = day5$dat5geneInfo,
                     CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
cluster_mixedpop1 <- colData(mixedpop1)[,1]
cluster_mixedpop2 <- colData(mixedpop2)[,1]
c_selectID <- 2
LSOLDA_dat <- bootstrap_scGPS(nboots = 2, mixedpop1 = mixedpop1, mixedpop2 = mixedpop2, genes=genes,
                        listData =list(), cluster_mixedpop1 = cluster_mixedpop1,
                        cluster_mixedpop2 = cluster_mixedpop2, c_selectID = c_selectID)
enrichment_test <- annotate_scGPS(genes, pvalueCutoff=0.05, gene_symbol=TRUE,
                                  output_filename = "PathwayEnrichment.xlsx", output_path = NULL)
clusterProfiler::dotplot(enrichment_test, showCategory=15)

}
