% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{clustering_scGPS}
\alias{clustering_scGPS}
\title{HC clustering for a number of resolutions}
\usage{
clustering_scGPS(object = NULL, ngenes = 1500, windows = seq(0.025:1, by =
  0.025), remove_outlier = c(0), nRounds = 1, PCA = TRUE, nPCs = 10)
}
\arguments{
\item{ngenes}{number of top variable genes to be used}

\item{remove_outlier}{a vector containing IDs for clusters to be removed
the default vector contains 0, as 0 is the cluster with singletons}

\item{nRounds}{number of iterations to remove a selected clusters}

\item{nPCs}{number of principal components from PCA dimensional reduction to be used}

\item{mixedpop1}{is a \linkS4class{SingleCellExperiment} object from the
train mixed population}
}
\value{
clustering results
}
\description{
performs 40 clustering runs or more depending on windows
}
\examples{
day5 <- sample2
mixedpop2 <-NewscGPS_SME(ExpressionMatrix = day5$dat5_counts, GeneMetadata = day5$dat5geneInfo,
                         CellMetadata = day5$dat5_clusters)
test <-clustering_scGPS(mixedpop2, remove_outlier = c(1))
}
\author{
Quan Nguyen, 2017-11-25
}
