% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{findMarkers_scGPS}
\alias{findMarkers_scGPS}
\title{find DE genes}
\usage{
findMarkers_scGPS(expression_matrix = NULL, cluster = NULL)
}
\arguments{
\item{expression_matrix}{is  a normalised expression matrix.}

\item{cluster}{corresponding cluster information in the expression_matrix
by going through CORE or from other method).}
}
\value{
a \code{matrix} with Eucleadean distance used for clustreting
}
\description{
Calculate distance matrix and perform hclust
}
\author{
Quan Nguyen, 2017-11-25
}
