% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{findMarkers_scGPS}
\alias{findMarkers_scGPS}
\title{find marker genes}
\usage{
findMarkers_scGPS(expression_matrix = NULL, cluster = NULL,
  selected_cluster = NULL, fitType = "local",
  dispersion_method = "per-condition")
}
\arguments{
\item{expression_matrix}{is  a normalised expression matrix.}

\item{cluster}{corresponding cluster information in the expression_matrix
by running CORE clustering or using other methods.}

\item{selected_cluster}{a vector of unique cluster ids to calculate}

\item{fitType}{string specifying 'local' or 'parametric' for DEseq dispersion
estimation}

\item{dispersion_method}{one of the options c( 'pooled', 'pooled-CR', 
per-condition', 'blind' )}
}
\value{
a \code{list} containing sorted DESeq analysis results
}
\description{
Find DE genes from comparing one clust vs remaining
}
\examples{
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, 
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
# depending on the data, the DESeq::estimateDispersions function requires
# suitable fitType
# and dispersion_method options
DEgenes <- findMarkers_scGPS(expression_matrix=assay(mixedpop1),
                        cluster = colData(mixedpop1)[,1],
                        selected_cluster=c(1,2),
                        fitType = "parametric", 
                        dispersion_method = "blind"
                        )
names(DEgenes)
}
\author{
Quan Nguyen, 2017-11-25
}
