% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{plotReduced_scGPS}
\alias{plotReduced_scGPS}
\title{plot reduced data}
\usage{
plotReduced_scGPS(reduced_dat, color_fac = factor(Sample_id),
  dims = c(1, 2), dimNames = c("Dim 1", "Dim 2"), palletes = NULL,
  legend_title = "Cluster")
}
\arguments{
\item{reduced_dat}{is a matrix with genes in rows and cells in columns}
}
\value{
a matrix with the top 20 CIDR dimensions
}
\description{
plot PCA, tSNE, and CIDR reduced datasets
}
\examples{
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, GeneMetadata = day2$dat2geneInfo,
                    CellMetadata = day2$dat2_clusters)
CIDR_dim <-CIDR_scGPS(expression.matrix=assay(mixedpop1))
p <-plotReduced_scGPS(CIDR_dim)
plot(p)
tSNE_dim <-tSNE_scGPS(expression.matrix=assay(mixedpop1))
p2 <-plotReduced_scGPS(tSNE_dim)
plot(p2)


}
