% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_Eucl_distance_NotPar}
\alias{rcpp_Eucl_distance_NotPar}
\title{Function to calculate Eucledean distance matrix without parallelisation}
\usage{
rcpp_Eucl_distance_NotPar(mat)
}
\arguments{
\item{mat}{an R matrix (expression matrix), with cells in rows and genes
in columns}
}
\value{
a distance matrix
}
\description{
Function to calculate Eucledean distance matrix without parallelisation
}
\examples{
mat_test <-matrix(rnbinom(100000,mu=0.01, size=10),nrow=1000)
#library(microbenchmark)
#microbenchmark(rcpp_Eucl_distance_NotPar(mat_test), dist(mat_test), times=3)

}
