% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_distance}
\alias{rcpp_parallel_distance}
\title{distance matrix using C++}
\usage{
rcpp_parallel_distance(mat)
}
\arguments{
\item{mat}{an R matrix (expression matrix), rows are genes, columns are cells}
}
\value{
a distance matrix
}
\description{
This function provides fast and memory efficient distance matrix calculation
}
\examples{
mat_test <-matrix(rnbinom(1000000,mu=0.01, size=10),nrow=10000)
#library(microbenchmark)
#microbenchmark(rcpp_parallel_distance(mat_test), dist(mat_test), times=3)

}
