% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_clustering.R
\name{SubClustering_scGPS}
\alias{SubClustering_scGPS}
\title{Subclustering for selected cells}
\usage{
SubClustering_scGPS(object = NULL, ngenes = 1500, windows = seq(0.025:1,
  by = 0.025), select_cell_index = NULL)
}
\arguments{
\item{object}{is a \linkS4class{SingleCellExperiment} object from the
train mixed population}

\item{select_cell_index}{a vector containing indexes for cells in selected clusters
to be reclustered}
}
\value{
clustering results
}
\description{
performs 40 clustering runs or more depending on windows
}
\author{
Quan Nguyen, 2018-01-31
}
