% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{topvar_scGPS}
\alias{topvar_scGPS}
\title{select top variable genes}
\usage{
topvar_scGPS(expression.matrix = NULL, ngenes = 1500)
}
\arguments{
\item{expression.matrix}{is a matrix with genes in rows and cells in columns}

\item{ngenes}{number of genes used for clustering calculations.}
}
\value{
a subsetted expression matrix with the top n most variable genes
}
\description{
subset a matrix by top variable genes
}
\examples{
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, 
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
SortedExprsMat <-topvar_scGPS(expression.matrix=assay(mixedpop1))
}
