% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{tSNE_scGPS}
\alias{tSNE_scGPS}
\title{tSNE}
\usage{
tSNE_scGPS(expression.matrix = NULL, ngenes = 1500, scaling = TRUE,
  theta = 0.5, perplexity = 30)
}
\arguments{
\item{expression.matrix}{An expression matrix, with genes in rows}

\item{ngenes}{number of genes used for clustering calculations.}

\item{scaling}{a logical of whether we want to scale the matrix}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy)}

\item{perplexity}{numeric; Perplexity parameter
(should not be bigger than 3 * perplexity < nrow(X) - 1, see details for 
interpretation)}
}
\value{
a tSNE reduced matrix containing three tSNE dimensions
}
\description{
calculate tSNE from top variable genes
}
\examples{
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts, 
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
t <-tSNE_scGPS(expression.matrix=assay(mixedpop1))

}
